Update WS sample
