+++ src/Middleware/WebSockets/samples/EchoApp/appsettings.json
@@ -1,0 +1,8 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Debug",
+      "Microsoft.AspNetCore": "Trace"
+    }
+  }
+}
